requirements.txt
Flask==0.12
Flask-RESTful==0.3.5
crochet==1.6.0
docker-py==1.10.6
M2Crypto==0.25.1
flask-marshmallow==0.7.0
blinker==1.4
futures==3.0.5
gunicorn==19.7.1
click==6.7
PyMySQL==0.7.10
Flask-SQLAlchemy==2.2
marshmallow-sqlalchemy==0.13.0
Flask-Migrate==2.0.3
